requirements.txt
Django==2.1.7
django-excel==0.0.10
django-templated-mail==1.1.1
djangorestframework==3.9.1
djangorestframework-simplejwt==4.0.0
djoser==1.4.0
lml==0.0.9
pyexcel==0.5.11
pyexcel-io==0.5.14
pyexcel-webio==0.1.4
PyJWT==1.7.1
pytz==2018.9
texttable==1.6.1
